+++ test/libsolidity/syntaxTests/natspec/docstring_double_empty.sol
@@ -1,0 +1,7 @@
+contract C {
+  ///
+  ///
+  function vote(uint id) public {
+  }
+}
+// ----