+++ Changelog.md
@@ -30,7 +30,9 @@ Compiler Features:
 Bugfixes:
  * Optimizer: Fixed a bug in BlockDeDuplicator.
  * Type Checker: Disallow assignments to storage variables of type ``mapping``.
-
+ * NatSpec: DocString block is terminated when encountering an empty line.
+ * Scanner: Fix bug when two empty NatSpec comments lead to scanning past EOL.
+ * Code Generator: Trigger proper unimplemented errors on certain array copy operations.
 
 ### 0.6.8 (2020-05-14)